disabled auto-create index setting
